requirements.txt
torch==1.1.0
torchvision==0.3.0
matplotlib==3.2.1
pandas==1.0.3
pillow==7.1.2
sklearn==0.0
scikit-image==0.17.1
tensorflow-gpu==1.9.0
opencv-python==4.2.0.34
tqdm==4.46.0
scikit-learn==0.22.2.post1
pylint==2.5.2
